move more functions from main to a module
